requirements.txt
annotated-types==0.6.0
bleach==6.1.0
certifi==2023.7.22
charset-normalizer==3.3.1
et-xmlfile==1.1.0
idna==3.4
kaggle==1.5.16
numpy==1.26.1
openpyxl==3.1.2
pandas==2.1.2
Pillow==10.1.0
pydantic==2.6.1
pydantic_core==2.16.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.1
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
text-unidecode==1.3
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
webencodings==0.5.1
